clarify table on separate components
