+++ src/assets/scss/feed.scss
@@ -204,6 +204,11 @@
       }
     }
 
+    .loading-table {
+      display: none;
+      appearance: none;
+    }
+
     .pagination-wrapper {
       @include typo-input;
