+++ src/components/CustomTable/index.tsx
@@ -1,0 +1,3 @@
+import Table from './CustomTable';
+
+export default Table;
