+++ src/components/MultiSelect.tsx
@@ -1,11 +1,12 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-// import { useState, useRef } from 'react';
 import Select, { components } from 'react-select';
 
-import { Option } from './Filter';
+// models
+import { Option } from '../interfaces/table-state';
+
 
 interface MultiSelect {
-  options: any,
+  options: Option[],
   value?: Option[],
   onChange?: any,
   isAllSelected?: boolean,
@@ -67,8 +68,6 @@ const MultiSelect = ({options, width, widthSelect, name, value, onChange, isAllS
     options.length && options?.filter(({ value }: Option) =>
       currOptions?.some((filterVal) => value === filterVal.value));
 
-
-  // const filteredOptions = filterOptions(options, value);
   const filteredSelectedOptions = value ? filterOptions(options, value) : [];
   const Option = (props: any) => (
     <>
